+++ src/osstask/transfer_oss2local.rs
@@ -117,7 +117,7 @@ impl TransferTaskActions for TransferOss2Local {
         &self,
         last_modify_timestamp: Option<i64>,
         object_list_file: &str,
-    ) -> Result<usize> {
+    ) -> Result<u64> {
         let client_source = self.source.gen_oss_client()?;
 
         // 若为持续同步模式，且 last_modify_timestamp 大于 0，则将 last_modify 属性大于last_modify_timestamp变量的对象加入执行列表
